Fixing agent compilation errors
